// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const StockStatus = {
  IN_STOCK: 'IN_STOCK',
  OUT_OF_STOCK: 'OUT_OF_STOCK',
  PREORDER: 'PREORDER',
  UNKNOWN: 'UNKNOWN'
} as const

export type StockStatus = (typeof StockStatus)[keyof typeof StockStatus]


export const DealStatus = {
  ACTIVE: 'ACTIVE',
  EXPIRED: 'EXPIRED',
  HIDDEN: 'HIDDEN',
  DRAFT: 'DRAFT'
} as const

export type DealStatus = (typeof DealStatus)[keyof typeof DealStatus]


export const SourceType = {
  CSV: 'CSV',
  SHEET: 'SHEET',
  SCRAPER: 'SCRAPER',
  MANUAL: 'MANUAL'
} as const

export type SourceType = (typeof SourceType)[keyof typeof SourceType]


export const PipelineRunStatus = {
  RUNNING: 'RUNNING',
  SUCCESS: 'SUCCESS',
  PARTIAL_SUCCESS: 'PARTIAL_SUCCESS',
  FAILED: 'FAILED'
} as const

export type PipelineRunStatus = (typeof PipelineRunStatus)[keyof typeof PipelineRunStatus]


export const IntakeType = {
  SCRAPED: 'SCRAPED',
  MANUAL: 'MANUAL',
  API: 'API'
} as const

export type IntakeType = (typeof IntakeType)[keyof typeof IntakeType]